// part9/patientor-frontend/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: Gender;
  ssn?: string;
  dateOfBirth?: string;
  entries: Entry[];
}

export interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export interface HospitalEntry extends BaseEntry {
    type: "Hospital";
    discharge: {
        date: string;
        criteria: string;
    }
}

export interface OccupationalHealthcareEntry extends BaseEntry {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: {
        startDate: string,
        endDate: string,
    }
}

export interface NewEntryAllInput extends BaseEntry {
  type: string;
  healthCheckRating: HealthCheckRating;
  discharge: {
      date: string;
      criteria: string;
  };
  employerName: string;
  sickLeave?: {
      startDate: string,
      endDate: string,
  }
}

export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;

// Define special omit for unions
export type UnionOmit
<T, K extends string | number | symbol> = 
  T extends unknown ? 
         Omit<T, K> : 
         never;

// Define Entry without the 'id' property
export type NewEntry = UnionOmit<Entry, 'id'>;


export enum EntryType {
  HospitalEntryType = 'Hospital',
  HealthCheckEntryType = 'HealthCheck',
  OccupationalHealthcareEntryType = 'OccupationalHealthcare'
}